Fri, Oct 28, 2022, 12:33 AM +00:00

+++ pset4/filter/helpers.c
@@ -105,10 +105,10 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            int count = 0;
-            float somaRed = 0;
-            float somaGreen = 0;
-            float somaBlue = 0;
+            float count = 0;
+            int somaRed = 0;
+            int somaGreen = 0;
+            int somaBlue = 0;
             for (int l = i-1; l <= i+1; l++)
             {
                 for (int t = j-1; t <= j+1; t++)
@@ -122,9 +122,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                     }
                 }
             }
-            image[i][j].rgbtRed = (int)round(somaRed / count);
-            image[i][j].rgbtGreen = (int)round(somaGreen / count);
-            image[i][j].rgbtBlue = (int)round(somaBlue / count);
+            image[i][j].rgbtRed = round(somaRed / count);
+            image[i][j].rgbtGreen = round(somaGreen / count);
+            image[i][j].rgbtBlue = round(somaBlue / count);
             }
         }
     return;
